% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_linear.R
\name{BS_rel}
\alias{BS_rel}
\title{correct t statistics from bootstrap approach2}
\usage{
BS_rel(sim_dat_tv_nested, ss)
}
\arguments{
\item{sim_dat_tv_nested}{a nested data frame with model fit statistics from IAP and Inf model}

\item{ss}{data size in each training, testing, and validation set.}
}
\value{
\item{tidy_table}{bias corrected statistics table from derivation approach}
}
\description{
get corrected t statistics from bootstrap approach: bootstrap from validation set using relationship model between real and predicted outcome
}
\examples{
\dontrun{
t_bs_real = BS_test(sim_dat_tv_nested)
}
}
\keyword{approach2}
\keyword{bootstrap}
\keyword{corrected}
\keyword{from}
\keyword{provide}
\keyword{statistics}
\keyword{t}
