% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_linear.R
\name{Deriv_BC}
\alias{Deriv_BC}
\title{correct t statistics from derivation approach}
\usage{
Deriv_BC(sim_dat_tv_nested, ss)
}
\arguments{
\item{sim_dat_tv_nested}{a nested data frame with model fit statistics from IAP and Inf model}

\item{ss}{data size in each training, testing, and validation set.}
}
\value{
\item{tidy_table}{bias corrected statistics table from derivation approach}
}
\description{
get corrected t statistics from derivation approach
}
\examples{
\dontrun{
t_pstar = Deriv_BC(sim_dat_tv_nested)
}
}
\keyword{approach}
\keyword{corrected}
\keyword{derivation}
\keyword{from}
\keyword{provide}
\keyword{statistics}
\keyword{t}
