% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_linear.R
\name{fit_IAP_Inf}
\alias{fit_IAP_Inf}
\title{fit Inference after prediction model and inference model}
\usage{
fit_IAP_Inf(sim_dat_tv)
}
\arguments{
\item{sim_dat_tv}{simulated data frame from iap_sim(ss, n_sim, beta1, beta2, beta3, beta4, method)}
}
\value{
\item{sim_dat_tv_nested}{provide a nested data frame with model fit statistics from IAP and Inf model}
}
\description{
fit Inference after prediction model and inference model
}
\examples{
\dontrun{
sim_dat_tv_nested = fit_IAP_Inf(sim_dat_tv)
}
}
\keyword{IAP}
\keyword{Inf}
\keyword{and}
\keyword{fit}
\keyword{model}
\keyword{provide}
\keyword{results}
