% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_linear.R
\name{IAP_model}
\alias{IAP_model}
\title{inference after prediction model}
\usage{
IAP_model(outcome, predictor)
}
\arguments{
\item{outcome}{input predicted outcome}

\item{predictor}{input predictor of interest}
}
\value{
\item{model}{provide inference after prediction model}
}
\description{
specify inference after prediction model using predicted outcome and predictor of interest
}
\examples{
\dontrun{
model = IAP_model(outcome, predictor)
}
}
\keyword{IAP}
\keyword{model}
\keyword{provide}
