% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_linear.R
\name{iap_sim}
\alias{iap_sim}
\title{Bias correction in statistical inference after predictions}
\usage{
iap_sim(ss, n_sim, beta1, beta2, beta3, beta4, method)
}
\arguments{
\item{ss}{data size in each training, testing, and validation set.}

\item{n_sim}{number of simulations}

\item{beta1}{coefficient for the first predictor}

\item{beta2}{coefficient for the second predictor}

\item{beta3}{coefficient for the third predictor}

\item{beta4}{coefficient for the fourth predictor}

\item{method}{indicate prediction method}
}
\value{
\item{sim_dat_tv}{data frame for simulated outcome, predictors, and predicted outcomes using specified method in training, testing, and validation set}
}
\description{
This function is used for simulating outcome and predictors from linear case, then using machine learning method to make predictions to get predicted outcome.
}
\examples{
\dontrun{
sim_dat_tv = iap_sim(ss, n_sim, beta1, beta2, beta3, beta4, method)
}
}
\keyword{data}
\keyword{frame}
\keyword{provide}
