% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_nonlinear.R
\name{iap_sim_nonlinear}
\alias{iap_sim_nonlinear}
\title{simulate categorical data frame}
\usage{
iap_sim_nonlinear(ss, n_sim, sim.test, method)
}
\arguments{
\item{ss}{data size in each training, testing, and validation set.}

\item{sim.test}{simulated verbal autopsy data}

\item{method}{specify prediction method}

\item{n.sim}{number of simulations}
}
\value{
\item{sim.test}{data frame for simulated outcome, predictors, and predicted outcomes using specified method in training, testing, and validation set}
}
\description{
This function is used for simulating outcome and predictors from nonlinear case, then using machine learning method to make predictions to get predicted outcome.
}
\examples{
\dontrun{
sim_dat_tv = iap_sim_nonlinear(ss, n_sim, sim.test, method)
}
}
\keyword{data}
\keyword{frame}
\keyword{provide}
