% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_linear.R
\name{Inf_model}
\alias{Inf_model}
\title{inference model}
\usage{
Inf_model(outcome, predictor)
}
\arguments{
\item{outcome}{input real outcome}

\item{predictor}{input predictor of interest}
}
\value{
\item{model}{provide inference model}
}
\description{
specify inference model between real outcome and predictor of interest
}
\examples{
\dontrun{
model = Inf_model(outcome, predictor)
}
}
\keyword{inference}
\keyword{model}
\keyword{provide}
