% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_func.R
\name{LevelTrans}
\alias{LevelTrans}
\title{Translate alphebatic matrix into numeric order matrix}
\usage{
LevelTrans(x, problevel)
}
\arguments{
\item{x}{alphabetic matrix}

\item{problevel}{customized probbase level}
}
\value{
\item{indic}{provide indicator}
 \item{Yt}{provide Yt}
 \item{cond.prob}{provide conditional probability matrix}
}
\description{
TTranslate alphebatic matrix into numeric order matrix
}
\examples{
\dontrun{
Trans_matrix = LevelTrans(x, problevel)
}
}
\keyword{matrix}
\keyword{numeric}
\keyword{order}
\keyword{provide}
