% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpi.R
\name{postpi}
\alias{postpi}
\title{postpi function}
\usage{
postpi(valid_dat, rel_model, inf_formula, bs = 100, seed = 1234)
}
\arguments{
\item{valid_dat}{validation set that contains predicted outcomes and covariates}

\item{rel_model}{model object created by `postpi_relate()`}

\item{inf_formula}{inference formula, eg. y ~ x1}

\item{bs}{number of bootstrap times. The default value is 100 times.}

\item{seed}{seed number}
}
\description{
This function provides the corrected inference result using a bootstrap approach
}
