% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpi_covariate_calib.R
\name{postpi_covariate_calibrate}
\alias{postpi_covariate_calibrate}
\title{postpi_covariate_calibrate function fits a model relating a true binary variable to its predicted probability, and other variables to be used in a downstream analysis.}
\usage{
postpi_covariate_calibrate(
  calib_data,
  true_var,
  pred_prob_var,
  covariates,
  outcome
)
}
\arguments{
\item{calib_data}{dataframe that contains the below variables needed for post-prediction inference calibration}

\item{true_var}{character string; column name of the true binary variable of interest in \code{calib_data}}

\item{pred_prob_var}{character string; column name of the predicted probability variable for the binary variable of interest in \\code{calib_data}}

\item{covariates}{character vector; column names of the covariates for the analysis of interest in \code{calib_data}}

\item{outcome}{character string; column name (or names in formula syntax) of the outcome variable for the analysis of interest in \code{calib_data}}
}
\value{
list containing glmnet object, optimal lambda parameter, and calibration formula.
}
\description{
This function takes in a labeled dataset for calibration containing:
(1) true measurements of a binary variable that is being predicted using ML
(2) predicted probabilities for this binary variable to be calibrated
(3) additional covariates to be used in a downstream analytic model
(4) the outcome of interest in the downstream analytic model
These variable names are specified as character or character vector inputs to the function, which are then combined into a model specification using standard R forumla syntax.
Note that in the case of a survival outcome, we recommend accounting for both the observed time and censoring indictor by specifying \code{time * status}.
}
\details{
The calibration model is fit as an l2-penalized logistic regression model, using the glmnet package.
The penalty parameter is set using cross-validation to minimize logistic deviance.
}
\examples{

data(EHRdata, package="postpi")
library(tidyverse)
library(survival)

labeled_data <- EHRdata[[1]]
postpi_calib <- postpi_covariate_calibrate(calib_data = labeled_data,
                                           true_var = "met_true",
                                           pred_prob_var = "met_prob",
                                           covariates = c("age_dx", "sex"),
                                           outcome = "time * status")

}
