% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpi_der.R
\name{postpi_der}
\alias{postpi_der}
\title{postpi_der function}
\usage{
postpi_der(test_dat, yobs, ypred, valid_dat, inf_formula)
}
\arguments{
\item{test_dat}{testing set that contains observed outcomes and predicted outcomes (continuous data) or probabilities of predicted outcomes (categorical data)}

\item{yobs}{name of the continuous observed outcome}

\item{ypred}{name of the continuous predicted outcome}

\item{valid_dat}{validation set that contains predicted outcomes and covariates}

\item{inf_formula}{inference formula, eg. y ~ x1}
}
\description{
This function provides the corrected inference result using a derivation approach for continuous data
}
