% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpi_relate.R
\name{postpi_relate}
\alias{postpi_relate}
\title{postpi_relate function}
\usage{
postpi_relate(test_dat, yobs, method = "knn")
}
\arguments{
\item{test_dat}{testing set that contains observed outcomes and predicted outcomes (continuous data) or probabilities of predicted outcomes (categorical data)}

\item{yobs}{name of the observed outcome in the testing set}
}
\description{
This function models the relationship between the observed and predicted outcomes.
}
