% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_linear.R
\name{realpred_model}
\alias{realpred_model}
\title{relationship model between real and predicted outcome}
\usage{
realpred_model(outcome, predictor)
}
\arguments{
\item{outcome}{input real outcome}

\item{predictor}{input predicted outcome}
}
\value{
\item{model}{provide relationship model between real and predicted outcome}
}
\description{
specify relationship model between real outcome and predicted outcome
}
\examples{
\dontrun{
model = realpred_model(outcome, predictor)
}
}
\keyword{model}
\keyword{provide}
\keyword{relationship}
