% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RINdata}
\alias{RINdata}
\title{Gene expression data for RNA quality prediction example}
\format{
A data frame with 4,769 rows and 202 variables:
\describe{
  \item{predictions}{RNA quality predictions}
  \item{actual}{RNA quality actual values}
  \item{region_x}{Gene expression data for region x}
  ...
}
}
\source{
\url{https://academic.oup.com/nar/article/46/9/e54/4920847}
}
\usage{
RINdata
}
\description{
A dataset containing data for 4769 samples with
gene expression information, predictions, and actual
values for RIN (a measure of RNA-quality)
}
\keyword{datasets}
