% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_linear.R
\name{t_IAP_Inf}
\alias{t_IAP_Inf}
\title{get t statistics from inference after prediction model and inference model}
\usage{
t_IAP_Inf(sim_dat_tv_nested)
}
\arguments{
\item{sim_dat_tv_nested}{a nested data frame with model fit statistics from IAP and Inf model}
}
\value{
\item{df}{a data frame with 2 colomns, each colomn with t statistics from IAP and Inf model}
}
\description{
get t statistics from necessary models
}
\examples{
\dontrun{
t_df = t_IAP_Inf(sim_dat_tv_nested)
}
}
\keyword{IAP}
\keyword{Inf}
\keyword{and}
\keyword{fit}
\keyword{model}
\keyword{provide}
\keyword{statistics}
\keyword{t}
