% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TISSUEdata}
\alias{TISSUEdata}
\title{Gene expression data for tissue prediction example}
\format{
A data frame with 288 rows and 2,285 variables:
\describe{
  \item{predictions}{Tissue predictions}
  \item{actual}{Actual tissue types}
  \item{region_x}{Gene expression data for region x}
  ...
}
}
\source{
\url{https://academic.oup.com/nar/article/46/9/e54/4920847}
}
\usage{
TISSUEdata
}
\description{
A dataset containing data for 288 samples with
gene expression information, predictions, and actual
values for tissue type.
}
\keyword{datasets}
