% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_nonlinear.R
\name{VA_data}
\alias{VA_data}
\title{Bias correction in statistical inference after predictions for categorical case}
\usage{
VA_data(n.sim)
}
\arguments{
\item{n.sim}{number of simulations}
}
\value{
\item{sim.test}{data frame for simulated outcome, predictors, and predicted outcomes using specified method in training, testing, and validation set}
}
\description{
This function is used for generating categorical verbal autopsy data
}
\examples{
\dontrun{
sim.test = VA_data(n.sim)
}
}
\keyword{data}
\keyword{frame}
\keyword{provide}
