% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_linear.R
\name{ypy_plot}
\alias{ypy_plot}
\title{plot relationship between outcome and predictor of interest, and relationship between outcome and predicted outcome}
\usage{
ypy_plot(sim_dat_tv)
}
\arguments{
\item{sim_dat_tv}{simulated data frame from iap_sim(ss, n_sim, beta1, beta2, beta3, beta4, method)}
}
\value{
\item{y_yp}{relationship plot between outcome and predicted outcome}
 \item{y_x}{relationship plot between outcome and predictor of interest}
}
\description{
This function is used for making plots between outcome and predictor of interest, and between outcome and predicted outcome
}
\examples{
\dontrun{
ypy_plot(sim_dat_tv)
}
}
\keyword{plot}
\keyword{provide}
\keyword{relationship}
