% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpi_relate.R
\name{postpi_relate}
\alias{postpi_relate}
\title{postpi_relate function models the the relationship between the observed and predicted outcomes.}
\usage{
postpi_relate(test_dat, yobs, method = "knn")
}
\arguments{
\item{test_dat}{testing set that contains observed outcomes and predicted outcomes (continuous data) or probabilities of predicted outcomes (categorical data)}

\item{yobs}{name of the observed outcome in the testing set}

\item{yobs}{name of the observed outcome in the testing set}
}
\value{
rel_model relationship model between observed outcomes and predicted outcomes/probabilities
}
\description{
This function takes in a data set (i.e. testing set) containing only observed and predicted outcomes.
It relates observed and predicted continuous outcomes through a gamma function.
For categorical outcomes, user inputs observed outcomes and the probabilities of predicted outcomes (i.e. probabilities for each predicted category).
It related categorical data through a user defined maching learning method from the caret package. The default method is k-nearest neighbours.
}
